Code cleaning

+++ libraries/keyboardmanager/keyboardmanager.cpp
@@ -2,13 +2,13 @@
 
 void KEYBD_read() {
 
-  byte data = (byte) ((GPIOC->IDR & 0x0010) >> 4) + ((GPIOC->IDR & 0x0020) >> 4) + ((GPIOB->IDR & 0x0001) << 2) + ((GPIOB->IDR & 0x0002) << 2) + ((GPIOC->IDR & 0x0008) << 1) + ((GPIOC->IDR & 0x0004) << 3) + ((GPIOC->IDR & 0x0001) << 6) + ((GPIOA->IDR & 0x0001) << 7)
+  byte data = ((GPIOC->IDR & 0x0010) >> 4) + ((GPIOC->IDR & 0x0020) >> 4) + ((GPIOB->IDR & 0x0001) << 2) + ((GPIOB->IDR & 0x0002) << 2) + ((GPIOC->IDR & 0x0008) << 1) + ((GPIOC->IDR & 0x0004) << 3) + ((GPIOC->IDR & 0x0001) << 6) + ((GPIOA->IDR & 0x0001) << 7)
 
-  GPIOJ->ODR = (unsigned short) (((data & 0x00FE) >> 1) | (GPIOJ->ODR & ODRPORTJMask));
-  GPIOG->ODR = (unsigned short) (((data & 0x0001) << 13) | (GPIOG->ODR & ODRPORTGMask));
+  GPIOJ->ODR = (((data & 0x00FE) >> 1) | (GPIOJ->ODR & ODRPORTJMask));
+  GPIOG->ODR = (((data & 0x0001) << 13) | (GPIOG->ODR & ODRPORTGMask));
 
-  GPIOA->ODR = (unsigned short) GPIOA->ODR | 0x0080;
-  GPIOA->ODR = (unsigned short) GPIOA->ODR & 0xFF7F;
+  GPIOA->ODR = GPIOA->ODR | 0x0080;
+  GPIOA->ODR = GPIOA->ODR & 0xFF7F;
 
 }
