+++ libraries/zpmanager/zpmanager.cpp
@@ -5,19 +5,19 @@ byte zeropage[256];
 // (address >= 0x0000) && (address <= 0x00FF)
 void ZP_read() {
 
-  unsigned int address = (unsigned int) (GPIOJ->IDR & 0x0080) + (GPIOE->IDR & 0x0040) + ((GPIOI->IDR & 0x8000) >> 10) + ((GPIOG->IDR & 0x0400) >> 6) + ((GPIOH->IDR & 0x8000) >> 12) + ((GPIOK->IDR & 0x0001) << 2) + ((GPIOI->IDR & 0x0800) >> 10) + ((GPIOK->IDR & 0x0004) >> 2);
+  uint16_t address = (GPIOJ->IDR & 0x0080) + (GPIOE->IDR & 0x0040) + ((GPIOI->IDR & 0x8000) >> 10) + ((GPIOG->IDR & 0x0400) >> 6) + ((GPIOH->IDR & 0x8000) >> 12) + ((GPIOK->IDR & 0x0001) << 2) + ((GPIOI->IDR & 0x0800) >> 10) + ((GPIOK->IDR & 0x0004) >> 2);
 
-  GPIOJ->ODR = (unsigned short) (((zeropage[address] & 0x00FE) >> 1) | (GPIOJ->ODR & ODRPORTJMask));
-  GPIOG->ODR = (unsigned short) (((zeropage[address] & 0x0001) << 13) | (GPIOG->ODR & ODRPORTGMask));
+  GPIOJ->ODR = (((zeropage[address] & 0x00FE) >> 1) | (GPIOJ->ODR & ODRPORTJMask));
+  GPIOG->ODR = (((zeropage[address] & 0x0001) << 13) | (GPIOG->ODR & ODRPORTGMask));
 
 }
 
 // (address >= 0x0000) && (address <= 0x00FF)
 void ZP_write() {
 
-  unsigned int address = (unsigned int) (GPIOJ->IDR & 0x0080) + (GPIOE->IDR & 0x0040) + ((GPIOI->IDR & 0x8000) >> 10) + ((GPIOG->IDR & 0x0400) >> 6) + ((GPIOH->IDR & 0x8000) >> 12) + ((GPIOK->IDR & 0x0001) << 2) + ((GPIOI->IDR & 0x0800) >> 10) + ((GPIOK->IDR & 0x0004) >> 2);
+  uint16_t address = (GPIOJ->IDR & 0x0080) + (GPIOE->IDR & 0x0040) + ((GPIOI->IDR & 0x8000) >> 10) + ((GPIOG->IDR & 0x0400) >> 6) + ((GPIOH->IDR & 0x8000) >> 12) + ((GPIOK->IDR & 0x0001) << 2) + ((GPIOI->IDR & 0x0800) >> 10) + ((GPIOK->IDR & 0x0004) >> 2);
 
-  zeropage[address] = (byte) (GPIOG->IDR & 0x0080) + ((GPIOE->IDR & 0x0030) << 1) + ((GPIOB->IDR & 0x0004) << 2) + ((GPIOI->IDR & 0x2000) >> 10) + ((GPIOI->IDR & 0x0400) >> 8) + ((GPIOK->IDR & 0x0080) >> 6) + ((GPIOI->IDR & 0x4000) >> 14);
+  zeropage[address] = (GPIOG->IDR & 0x0080) + ((GPIOE->IDR & 0x0030) << 1) + ((GPIOB->IDR & 0x0004) << 2) + ((GPIOI->IDR & 0x2000) >> 10) + ((GPIOI->IDR & 0x0400) >> 8) + ((GPIOK->IDR & 0x0080) >> 6) + ((GPIOI->IDR & 0x4000) >> 14);
 
 }
